Clear lrus in sync
After a clear, during a sync, the RO instance has to clear its lru as well, otherwise it will keep finding old values in the lru. To do this it check if the generation is updated in the pack IO.
Contents, commits and nodes stores each have their own lru but share the same instance of pack IO. Therefore the first store to sync, updates pack IO and has to clear the lrus for all the other stores. This is achieved with a [clear_lru] callback.